requirements.txt
anthropic>=0.75.0
openai>=1.0.0
python-dotenv>=1.0.0
google-genai>=0.3.0
xai-sdk
